Moved most variables to f64, trying to fix the exponential of the stopping distance formula. It did not work

+++ src/components.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use uom::si::f32 as quantities;  // Should I use f64?
+use uom::si::f64 as quantities;  // Should I use f64?
 
 #[derive(Component)]
 pub struct CenterOfMass;
@@ -10,7 +10,7 @@ pub struct SailElement;
 #[derive(Component, Debug)]
 pub struct ESail {
     pub elements:           Vec<Entity>,
-    pub resting_distance:   f32,    // meters
+    pub resting_distance:   f64,    // meters
     //pub resting_distance:   quantities::Length,
 }
 
@@ -27,9 +27,9 @@ pub struct ElectricallyCharged {
 // I could call this SailElement and make everything simpler
 #[derive(Component, Debug)]
 pub struct VerletObject {
-    pub previous_x: f32,    // Wondering if this should stay in pixels or become physical meters
-    pub previous_y: f32,
-    pub current_x:  f32,
-    pub current_y:  f32,
-    pub is_deployed: bool,  // This would be better in another component, SailElement maybe
+    pub previous_x:     f64,    // Wondering if this should stay in pixels or become physical meters
+    pub previous_y:     f64,
+    pub current_x:      f64,
+    pub current_y:      f64,
+    pub is_deployed:    bool,  // This would be better in another component, SailElement maybe
 }
